pyproject.toml
[project]
name = "fr8-ai-coach"
version = "0.1.0"
description = "Real-time AI coaching backend with video capture and voice feedback"
dependencies = [
    "opencv-python",
    "google-generativeai",
    "python-dotenv",
    "google-genai>=1.28.0",
    "openai>=1.98.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
